Revert "[Datasets] Update usages of `np.array` constructor for ragged ndarrays to specify `dtype=object` (#31544)" (#31660)
This reverts commit f07a1b7e9762708701fa83f1d0664191daa10b3c.
